requirements.txt
python-telegram-bot==20.6
pandas>=1.5
openpyxl
lxml
requests
tabulate
python-dotenv
